pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pycoder"
version = "0.1.0"
description = "A flexible AI agent management system with support for multiple LLM providers"
authors = [
    {name = "Yigal Weinberger", email = "yigalw10@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.103.1",
    "pydantic>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.5.0"]
gemini = ["google-generativeai>=0.3.0"]
groq = ["groq>=0.3.0"]
openrouter = ["openrouter>=0.2.0"]
all = ["openai>=1.0.0", "anthropic>=0.5.0", "google-generativeai>=0.3.0", "groq>=0.3.0", "openrouter>=0.2.0"]

[project.scripts]
pycoder = "pycoder.run_agents_server:run_cli"

[project.urls]
"Homepage" = "https://github.com/yigalw10/pycoder"
"Bug Tracker" = "https://github.com/yigalw10/pycoder/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
